% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download from GISAID}
\usage{
download(
  credentials,
  list_of_accession_ids,
  get_sequence = FALSE,
  clean_up = TRUE
)
}
\arguments{
\item{list_of_accession_ids}{list of accession_id from GISAID}

\item{get_sequence}{load sequences into data.frame after download}

\item{clean_up}{delete downloaded files (e.g. fasta files) after download}
}
\value{
data.frame of complete data
}
\description{
Download from GISAID
}
