% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to GISAID}
\usage{
login(username, password)
}
\arguments{
\item{username}{GISAID username.}

\item{password}{GISAID password.}
}
\value{
credentials used to query GISAID.
}
\description{
Login to GISAID
}
\examples{
login(username, password)
}
