% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query GISAID Database}
\usage{
query(
  credentials,
  location = NULL,
  lineage = NULL,
  from = NULL,
  from_subm = NULL,
  to = NULL,
  to_subm = NULL,
  start_index = 0,
  nrows = 50,
  load_all = FALSE,
  low_coverage_excl = FALSE,
  complete = FALSE,
  high_coverage = FALSE,
  collection_date_complete = FALSE,
  total = FALSE
)
}
\arguments{
\item{credentials}{GISAID credentials.}

\item{location}{search for entries based on geographic location.}

\item{lineage}{search for entries based on pango lineage designations.}

\item{from}{search from specific collection date.}

\item{from_subm}{search from specific submission date.}

\item{to}{search to specific collection date.}

\item{to_subm}{search to specific submission date.}

\item{start_index}{page through results.}

\item{nrows}{number of results to return.}

\item{load_all}{return all results.}

\item{low_coverage_excl}{exclude low coverage entries from the results.}

\item{complete}{include only complete entries in the results.}

\item{high_coverage}{include only high coverage entries in the results.}

\item{collection_date_complete}{include only entries with complete in collection date the results.}

\item{total}{returns the total number of sequences matching the query.}
}
\value{
Dataframe.
}
\description{
Query GISAID Database
}
