% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA amino acids file into a dataframe}
\usage{
read_fasta(file = NULL, get_sequence = TRUE)
}
\description{
This function reads a FASTA amino acids file into a dataframe
}
\details{
TAKEN FROM AMPIR: https://github.com/Legana/ampir
}
